+++ src/maasserver/api/raid.py
@@ -17,8 +17,12 @@ from maasserver.api.support import OperationsHandler
 from maasserver.enum import (
     FILESYSTEM_TYPE,
     NODE_PERMISSION,
+    NODE_STATUS,
 )
-from maasserver.exceptions import MAASAPIValidationError
+from maasserver.exceptions import (
+    MAASAPIValidationError,
+    NodeStateViolation,
+)
 from maasserver.forms import (
     CreateRaidForm,
     UpdateRaidForm,
@@ -67,9 +71,13 @@ class RaidsHandler(OperationsHandler):
         :param spare_partitions: Spare partitions to add to the RAID.
 
         Returns 404 if the node is not found.
+        Returns 409 if the node is not Ready.
         """
         node = Node.nodes.get_node_or_404(
-            system_id, request.user, NODE_PERMISSION.EDIT)
+            system_id, request.user, NODE_PERMISSION.ADMIN)
+        if node.status != NODE_STATUS.READY:
+            raise NodeStateViolation(
+                "Cannot create RAID because the node is not Ready.")
         form = CreateRaidForm(node, data=request.data)
         if form.is_valid():
             return form.save()
@@ -168,9 +176,14 @@ class RaidHandler(OperationsHandler):
                RAID.
 
         Returns 404 if the node or RAID is not found.
+        Returns 409 if the node is not Ready.
         """
         raid = RAID.objects.get_object_or_404(
-            system_id, raid_id, request.user, NODE_PERMISSION.EDIT)
+            system_id, raid_id, request.user, NODE_PERMISSION.ADMIN)
+        node = raid.get_node()
+        if node.status != NODE_STATUS.READY:
+            raise NodeStateViolation(
+                "Cannot update RAID because the node is not Ready.")
         form = UpdateRaidForm(raid, data=request.data)
         if form.is_valid():
             return form.save()
@@ -181,8 +194,13 @@ class RaidHandler(OperationsHandler):
         """Delete RAID on node.
 
         Returns 404 if the node or RAID is not found.
+        Returns 409 if the node is not Ready.
         """
         raid = RAID.objects.get_object_or_404(
-            system_id, raid_id, request.user, NODE_PERMISSION.EDIT)
+            system_id, raid_id, request.user, NODE_PERMISSION.ADMIN)
+        node = raid.get_node()
+        if node.status != NODE_STATUS.READY:
+            raise NodeStateViolation(
+                "Cannot delete RAID because the node is not Ready.")
         raid.delete()
         return rc.DELETED
